# Conflicts:
#   frontend/src/pages/user/test.tsx

+++ backend/src/main/java/com/ssafy/fullerting/exArticle/service/ExArticleService.java
@@ -319,6 +319,13 @@ public class ExArticleService {
 
         List<Image> image = imageRepository.findAllByExArticleId(ex_article_id);
 
+        UserResponse userResponse = userService.getUserInfo();
+        CustomUser user = userResponse.toEntity(userResponse);
+
+        if (article.getUser().getId() != user.getId()) {
+            throw new ExArticleException(ExArticleErrorCode.NOT_MINE);
+        }
+
         for (Image image1 : image) {
             imageRepository.delete(image1);
             amazonS3Service.deleteFile(image1.getImgStoreUrl()); //s3
